requirements.txt
fastapi
asyncpg
python-dotenv
httpx
cryptography
uvicorn
anthropic
jinja2
boto3
bcrypt
python-jose[cryptography]
python-multipart
pypdf
reportlab
Pillow
